pyproject.toml
[project]
name = "io-helpers"
description = "A custom nodes package for ComfyUI that enhances workflow flexibility by providing specialized nodes for saving and loading intermediate data (encoded prompts and sampled latents) in multiple formats. This package leverages helper classes for file I/O, supports gzip compression for efficient storage, and integrates progress feedback via a progress bar to improve user experience during long operations."
version = "1.1.8"
readme = "README.md"
authors = [
    { name = "Austin Brown", email = "austinbrown34@gmail.com" }
]
license = { file = "LICENSE" }
keywords = [
    "comfyui",
    "custom-nodes",
    "prompt encoding",
    "latent sampling",
    "progress bar",
    "file I/O",
    "gzip"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "torch>=1.7",
    "numpy>=1.19",
    "safetensors>=0.1.0",
    "dill>=0.3.3",
    "requests",
    "aiohttp",
    "Pillow"
]

[project.urls]
Repository = "https://github.com/austinbrown34/ComfyUI-IO-Helpers"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "austinbrown34"
DisplayName = "ComfyUI-IO-Helpers"
Icon = "https://comfyui-io-helpers.s3.us-east-1.amazonaws.com/comfyui-io-helpers-scaled-banner.png"
